Work on point in polygon

+++ src/main/scala/ComputationMethod.scala
@@ -1,7 +1,7 @@
 import org.apache.spark.sql.Dataset
 
 
-final case class Result(state: String, county: String, averageElevation: Float)
+final case class Result(state: String, county: String, averageElevation: Double)
 
 trait ComputationMethod {
   def compute(rasterData: Dataset[Pixel], vectorData: Dataset[Region]): Dataset[Result]
